+++ u_cells/layers/__init__.py
@@ -1,2 +1,4 @@
 from .CoordConv import CoordConv
 from .GradCAM import GradCAM
+from .ConvBlock import ConvBlock
+from .UpConvBlock import UpConvBlock
